Dyn4cast: Dynamic Modeling and Machine Learning Environment

Estimates, predict and forecast dynamic models as well as Machine Learning metrics which assists in model selection for further analysis. The package also have capabilities to provide tools and metrics that are useful in machine learning and modeling. For example, there is quick summary, percent sign, Mallow's Cp tools and others. The ecosystem of this package is analysis of economic data for national development. The package is so far stable and has high reliability and efficiency as well as time-saving. The package is a variety but the following references are important guide to the major themes in the package (Hyndman & Athanasopoulos (2014 ISBN 978-0-9875071-0-5), Alkire & Santos (2014, doi.org/10.1016/j.worlddev.2014.01.026)).

See also

Author

Maintainer: Job Nmadu job.nmadu@gmail.com (ORCID)

Authors:

Job Nmadu job.nmadu@gmail.com (ORCID)