Sequential Computation of Dynamic Multidimensional Poverty Indices (MDPI)

This function computes the indices and all associated measures of multidimensional poverty sequentially in a dynamic way. Sequentially the function computes Incidence of poverty (H = q / n), Adjusted incidence of poverty (H / (q / D)), Deprivation Score of each dimension in the computation, Intensity of poverty (A), Multidimensional poverty index (MDPI = H * A), the Contribution in % of each of the dimensions to MDPI, and Average deprivation among the deprived (A * D). Dynamically, it computes the various indices for between three and nine dimensions (D). The first five dimensions included in the computations are Health, Education, Living standard, Social security and, Employment and Income depending on the choice of the user. Four additional dimensions can be included in the computations. The computations are carried out either for the national sample data or can be dis-aggregated based on grouping factors, like region, sex, gender, marital status or any suitable one. The cut-off mark demarcating poor (q) and non-poor (n-q) members in the sample (n) is defaulted to 0.4 but can be varied as may be dictated by the interests or the need for the computation. The computations are in line with various procedures already outlined in literature starting with the work of Alkire et. al, (2015) but has been expanded from three dimensions to nine. Each dimension is given equal weight in the computation but all indicators are weighted in line with existing guidelines in Alkire & Foster (2011) and Alkire & Santos (2010). See also Alkire & Santos (2014) and Chan & Wong (2024).

mdpi(
  data,
  dm,
  Bar = 0.4,
  id_addn = NULL,
  Factor = NULL,
  plots = NULL,
  id = c("Health", "Education", "Living standard"),
  id_add = "Social security",
  id_add1 = "Employment and Income"
)

Arguments

data

data frame containing all the variables for the computation. Note that the variables to be used for the computation must be coded (0,1).

dm

list of vectors of indicators making up each dimension to be computed

Bar

an optional vector of cut-of used to divide the population into those in the poverty category and those that are not. Defaults to 0.4 if not supplied.

id_addn

an optional vector of additional dimensions to be used for the computation up to a maximum of four.

Factor

an optional grouping factor for the computation which must be a variable in the data. If not supplied, only the national MDPI will be computed.

plots

plots of the various measures. For this to be possible, the number of options in the Factor argument must be less than 41. The default is NULL. To produce the plots, any character string will overwrite the default.

id

a vector of the first three dimensions used in the computation given as Health, Education and Living standard. Can be redefined but must match the indicators and cannot be NULL.

id_add

a vector of the fourth dimension in the computation given as Social security. Can be re-defined but never NULL.

id_add1

a vector of the fifth dimension in the computation given as Employment and Income. Can be re-defined but never NULL.

Value

A list with the following components:

MDPI_p

Publication-ready table of the factor and national MDPI prepared with summarymodels package. Will not return if only national computation is carried out.

MDPI

Data frame of the factor and national MDPI. Will not return if only national computation is carried out.

MDPI mean

Data frame of the mean MDPI. Will not return if only national computation is carried out.

MDPI SD

Data frame of the SD of MDPI. Will not return if only national computation is carried out.

national

Data frame of national MDPI with mean and SD.

dimensions

Data frame of the scores for each dimension in the computation.

Score

Data frame of the scores for each indicator in the computation.

References

Alkire, S. & Foster, J. (2011). Counting and Multidimensional Poverty Measurement. Journal of Public Economics 95(7-8): 476–87. https://doi.org/10.1016/j.jpubeco.2010.11.006.

Alkire, S., Foster, J. E., Seth, S., Santos, M. E., Roche, J., & Ballon, P. (2015). Multidimensional poverty measurement and analysis. Oxford University Press.

Alkire, S. & Santos, M. E. (2010). Acute Multidimensional Poverty: A New Index for Developing Countries. Oxford Poverty and Human Development Initiative (OPHI) Working Paper No. 38.

Alkire, S. & Santos, M. E. (2014). Measuring Acute Poverty in the Developing World: Robustness and Scope of the Multidimensional Poverty Index. World Development 59:251-274. https://doi.org/10.1016/j.worlddev.2014.01.026.

Siu Ming Chan & Hung Wong (2024): Measurement and determinants of multidimensional poverty: the case of Hong Kong, Journal of Asian Public Policy, DOI: 10.1080/17516234.2024.2325857

Examples

# # Not run, uncomment to run
# # data from `MPI` package
# data <- mdpi1
# dm <- list(d1 = c("Child.Mortality", "Access.to.health.care"),
#            d2 = c("Years.of.education", "School.attendance", "School.lag"),
#            d3 = c("Cooking.Fuel", "Access.to.clean.source.of.water",
#                   "Access.to.an.improve.sanatation", "Electricity",
#                   "Housing.Materials", "Asset.ownership"))
# mdpi(data, dm, plots = "t", Factor = "Region")
# mdpi(data, dm, plots = "t")
#
# # data from `mpitbR` package
# data <- mdpi2
# dm <- list(d1 = c("d_nutr","d_cm"),
#            d2 = c("d_satt","d_educ"),
#            d3 = c("d_elct","d_sani","d_wtr","d_hsg","d_ckfl","d_asst"))
# mdpi(data, dm, plots = "t", Factor = "region")
# mdpi(data, dm, plots = "t")