Dyn4cast: Dynamic Modeling and Machine Learning Environment

Estimates, predict and forecast dynamic models as well as Machine Learning metrics which assists in model selection for further analysis. The package also have capabilities to provide tools and metrics that are useful in machine learning and modeling. For example, there is quick summary, percent sign, Mallow's Cp tools and others. The ecosystem of this package is analysis of economic data for national development. The package is so far stable and has high reliability and efficiency as well as time-saving.

See also

Author

Maintainer: Job Nmadu job.nmadu@gmail.com (ORCID)