Garrett Ranking of Categorical Data

There are three main types of ranking: Standard competition, Ordinal and Fractional. Garrett's Ranking Technique is the application of fractional ranking in which the data points are ordered and given an ordinal number/rank. The ordering and ranking provide additional information which may not be available from frequency distribution. Again, the ordering is based on the level of seriousness or severity of the data point from the view point of the respondent. Ranking enables ease of comparison and makes grouping more meaningful. It is used in social science, psychology and other survey types of research. This functions performs Garrett Ranking of up to 15 ranks.

Usage

garrett_ranking(data, num_rank, ranking = NULL, m_rank = c(2:15))

Arguments

data

The data for the Garrett Ranking, must be a data.frame.

num_rank

A vector representing the number of ranks applied to the data. If the data is a five-point Likert-type data, then number of ranks is 5.

ranking

A vector of list representing the ranks applied to the data. If not available, positional ranks are applied.

m_rank

The scope of the ranking methods which is between 2 and 15.

Value

A list with the following components:

Data mean table

Table of data ranked using simple average.

Garrett ranked data

Table of data ranked using Garrett mean score.

Garrett value

Table of ranking Garrett values

Examples

garrett_data <- data.frame(garrett_data)
ranking <- c("Serious constraint", "Constraint",
"Not certain it is a constraint", "Not a constraint",
"Not a serious constraint")

## ranking is supplied
garrett_ranking(garrett_data, 5, ranking)
#> New names:
#> • `` -> `...1`
#> • `` -> `...8`
#> New names:
#> • `` -> `...1`
#> $`Data mean table`
#>    S/No Description Mean Remark Rank
#> 1     1          S1 14.8  Above    1
#> 2     2          S2  8.2  Above    2
#> 3     7          S7  7.0  Above    3
#> 4    13         S13  7.0  Above    4
#> 5     3          S3  6.0  Above    5
#> 6     9          S9  4.5  Above    6
#> 7    15         S15  4.5  Above    7
#> 8     6          S6  4.0  Above    8
#> 9    12         S12  4.0  Above    9
#> 10    5          S5  3.4  Above   10
#> 11   11         S11  3.4  Above   11
#> 12    4          S4  3.3  Above   12
#> 13   10         S10  3.3  Above   13
#> 14    8          S8  1.9  Below   14
#> 15   14         S14  1.9  Below   15
#> 
#> $`Garrett value`
#> # A tibble: 5 × 4
#>   Number `Garrett point` `Garrett index` `Garrett value`
#>    <dbl>           <dbl>           <dbl>           <dbl>
#> 1      1            3.33              15              85
#> 2      2           10                 25              75
#> 3      3           16.7               31              69
#> 4      4           23.3               36              64
#> 5      5           30                 40              60
#> 
#> $`Garrett ranked data`
#>    S/No Description Serious constraint Constraint
#> 1     2          S2                  5          3
#> 2     9          S9                  7          6
#> 3    15         S15                  7          6
#> 4     5          S5                 10          2
#> 5    11         S11                 10          2
#> 6     4          S4                  4          4
#> 7    10         S10                  4          4
#> 8     3          S3                  1          2
#> 9     1          S1                  0          0
#> 10    6          S6                  0          4
#> 11   12         S12                  0          4
#> 12    7          S7                  0          2
#> 13   13         S13                  0          2
#> 14    8          S8                  0          0
#> 15   14         S14                  0          0
#>    Not certain it is a constraint Not a constraint Not a serious constraint
#> 1                               2                2                        1
#> 2                               0                5                        1
#> 3                               0                5                        1
#> 4                               8                5                        0
#> 5                               8                5                        0
#> 6                               6                7                        3
#> 7                               6                7                        3
#> 8                               5                5                        1
#> 9                               2                1                        0
#> 10                              6                5                        6
#> 11                              6                5                        6
#> 12                              0                2                        2
#> 13                              0                2                        2
#> 14                              5                2                       17
#> 15                              5                2                       17
#>    Total Total Garrett Score Mean score Rank
#> 1     13                 976         75    1
#> 2     19                1425         75    2
#> 3     19                1425         75    3
#> 4     25                1872         75    4
#> 5     25                1872         75    5
#> 6     24                1682         70    6
#> 7     24                1682         70    7
#> 8     14                 960         69    8
#> 9      3                 202         67    9
#> 10    21                1394         66   10
#> 11    21                1394         66   11
#> 12     6                 398         66   12
#> 13     6                 398         66   13
#> 14    24                1493         62   14
#> 15    24                1493         62   15
#> 

# ranking not supplied
garrett_ranking(garrett_data, 5)
#> New names:
#> • `` -> `...1`
#> • `` -> `...8`
#> New names:
#> • `` -> `...1`
#> $`Data mean table`
#>    S/No Description Mean Remark Rank
#> 1     1          S1 14.8  Above    1
#> 2     2          S2  8.2  Above    2
#> 3     7          S7  7.0  Above    3
#> 4    13         S13  7.0  Above    4
#> 5     3          S3  6.0  Above    5
#> 6     9          S9  4.5  Above    6
#> 7    15         S15  4.5  Above    7
#> 8     6          S6  4.0  Above    8
#> 9    12         S12  4.0  Above    9
#> 10    5          S5  3.4  Above   10
#> 11   11         S11  3.4  Above   11
#> 12    4          S4  3.3  Above   12
#> 13   10         S10  3.3  Above   13
#> 14    8          S8  1.9  Below   14
#> 15   14         S14  1.9  Below   15
#> 
#> $`Garrett value`
#> # A tibble: 5 × 4
#>   Number `Garrett point` `Garrett index` `Garrett value`
#>    <dbl>           <dbl>           <dbl>           <dbl>
#> 1      1            3.33              15              85
#> 2      2           10                 25              75
#> 3      3           16.7               31              69
#> 4      4           23.3               36              64
#> 5      5           30                 40              60
#> 
#> $`Garrett ranked data`
#>    S/No Description 1st Rank 2nd Rank 3rd Rank 4th Rank 5th Rank Total
#> 1     2          S2        5        3        2        2        1    13
#> 2     9          S9        7        6        0        5        1    19
#> 3    15         S15        7        6        0        5        1    19
#> 4     5          S5       10        2        8        5        0    25
#> 5    11         S11       10        2        8        5        0    25
#> 6     4          S4        4        4        6        7        3    24
#> 7    10         S10        4        4        6        7        3    24
#> 8     3          S3        1        2        5        5        1    14
#> 9     1          S1        0        0        2        1        0     3
#> 10    6          S6        0        4        6        5        6    21
#> 11   12         S12        0        4        6        5        6    21
#> 12    7          S7        0        2        0        2        2     6
#> 13   13         S13        0        2        0        2        2     6
#> 14    8          S8        0        0        5        2       17    24
#> 15   14         S14        0        0        5        2       17    24
#>    Total Garrett Score Mean score Rank
#> 1                  976         75    1
#> 2                 1425         75    2
#> 3                 1425         75    3
#> 4                 1872         75    4
#> 5                 1872         75    5
#> 6                 1682         70    6
#> 7                 1682         70    7
#> 8                  960         69    8
#> 9                  202         67    9
#> 10                1394         66   10
#> 11                1394         66   11
#> 12                 398         66   12
#> 13                 398         66   13
#> 14                1493         62   14
#> 15                1493         62   15
#> 

# you can rank subset of the data
garrett_ranking(garrett_data, 8)
#> New names:
#> • `` -> `...1`
#> • `` -> `...11`
#> New names:
#> • `` -> `...1`
#> $`Data mean table`
#>    S/No Description Mean Remark Rank
#> 1     1          S1 14.8  Above    1
#> 2     2          S2  8.2  Above    2
#> 3     7          S7  7.0  Above    3
#> 4    13         S13  7.0  Above    4
#> 5     3          S3  6.0  Above    5
#> 6     9          S9  4.5  Above    6
#> 7    15         S15  4.5  Above    7
#> 8     6          S6  4.0  Below    8
#> 9    12         S12  4.0  Below    9
#> 10    5          S5  3.4  Below   10
#> 11   11         S11  3.4  Below   11
#> 12    4          S4  3.3  Below   12
#> 13   10         S10  3.3  Below   13
#> 14    8          S8  1.9  Below   14
#> 15   14         S14  1.9  Below   15
#> 
#> $`Garrett value`
#> # A tibble: 8 × 4
#>   Number `Garrett point` `Garrett index` `Garrett value`
#>    <dbl>           <dbl>           <dbl>           <dbl>
#> 1      1            3.33              15              85
#> 2      2           10                 25              75
#> 3      3           16.7               31              69
#> 4      4           23.3               36              64
#> 5      5           30                 40              60
#> 6      6           36.7               43              57
#> 7      7           43.3               47              53
#> 8      8           50                 50              50
#> 
#> $`Garrett ranked data`
#>    S/No Description 1st Rank 2nd Rank 3rd Rank 4th Rank 5th Rank 6th Rank
#> 1     7          S7        4        2        2        0        2        0
#> 2    13         S13        4        2        2        0        2        0
#> 3     2          S2        2        0        2        5        3        2
#> 4     9          S9        0        4        4        7        6        0
#> 5    15         S15        0        4        4        7        6        0
#> 6     3          S3        1        3        4        1        2        5
#> 7     5          S5        0        1        0       10        2        8
#> 8    11         S11        0        1        0       10        2        8
#> 9     4          S4        0        1        3        4        4        6
#> 10   10         S10        0        1        3        4        4        6
#> 11    6          S6        0        1        1        0        4        6
#> 12   12         S12        0        1        1        0        4        6
#> 13    1          S1        0        0        0        0        0        2
#> 14    8          S8        1        0        0        0        0        5
#> 15   14         S14        1        0        0        0        0        5
#>    7th Rank 8th Rank Total Total Garrett Score Mean score Rank
#> 1         2        2    14                 954         68    1
#> 2         2        2    14                 954         68    2
#> 3         2        1    17                1078         63    3
#> 4         5        1    27                1699         63    4
#> 5         5        1    27                1699         63    5
#> 6         5        1    22                1370         62    6
#> 7         5        0    26                1556         60    7
#> 8         5        0    26                1556         60    8
#> 9         7        3    28                1641         59    9
#> 10        7        3    28                1641         59   10
#> 11        5        6    23                1291         56   11
#> 12        5        6    23                1291         56   12
#> 13        1        0     3                 167         56   13
#> 14        2       17    25                1326         53   14
#> 15        2       17    25                1326         53   15
#> 

garrett_ranking(garrett_data, 4)
#> New names:
#> • `` -> `...1`
#> • `` -> `...7`
#> New names:
#> • `` -> `...1`
#> $`Data mean table`
#>    S/No Description Mean Remark Rank
#> 1     1          S1 14.8  Above    1
#> 2     2          S2  8.2  Above    2
#> 3     7          S7  7.0  Above    3
#> 4    13         S13  7.0  Above    4
#> 5     3          S3  6.0  Above    5
#> 6     9          S9  4.5  Above    6
#> 7    15         S15  4.5  Above    7
#> 8     6          S6  4.0  Above    8
#> 9    12         S12  4.0  Above    9
#> 10    5          S5  3.4  Above   10
#> 11   11         S11  3.4  Above   11
#> 12    4          S4  3.3  Above   12
#> 13   10         S10  3.3  Above   13
#> 14    8          S8  1.9  Below   14
#> 15   14         S14  1.9  Below   15
#> 
#> $`Garrett value`
#> # A tibble: 4 × 4
#>   Number `Garrett point` `Garrett index` `Garrett value`
#>    <dbl>           <dbl>           <dbl>           <dbl>
#> 1      1            3.33              15              85
#> 2      2           10                 25              75
#> 3      3           16.7               31              69
#> 4      4           23.3               36              64
#> 
#> $`Garrett ranked data`
#>    S/No Description 1st Rank 2nd Rank 3rd Rank 4th Rank Total
#> 1     9          S9        6        0        5        1    12
#> 2    15         S15        6        0        5        1    12
#> 3     2          S2        3        2        2        1     8
#> 4     5          S5        2        8        5        0    15
#> 5    11         S11        2        8        5        0    15
#> 6     3          S3        2        5        5        1    13
#> 7     4          S4        4        6        7        3    20
#> 8    10         S10        4        6        7        3    20
#> 9     1          S1        0        2        1        0     3
#> 10    7          S7        2        0        2        2     6
#> 11   13         S13        2        0        2        2     6
#> 12    6          S6        4        6        5        6    21
#> 13   12         S12        4        6        5        6    21
#> 14    8          S8        0        5        2       17    24
#> 15   14         S14        0        5        2       17    24
#>    Total Garrett Score Mean score Rank
#> 1                  919         77    1
#> 2                  919         77    2
#> 3                  607         76    3
#> 4                 1115         74    4
#> 5                 1115         74    5
#> 6                  954         73    6
#> 7                 1465         73    7
#> 8                 1465         73    8
#> 9                  219         73    9
#> 10                 436         73   10
#> 11                 436         73   11
#> 12                1519         72   12
#> 13                1519         72   13
#> 14                1601         67   14
#> 15                1601         67   15
#>