Garrett Ranking of Categorical Data

There are three main types of ranking: Standard competition, Ordinal and Fractional. Garrett's Ranking Technique is the application of fractional ranking in which the data points are ordered and given an ordinal number/rank. The ordering and ranking provide additional information which may not be available from frequency distribution. Again, the ordering is based on the level of seriousness or severity of the data point from the view point of the respondent. Ranking enables ease of comparison and makes grouping more meaningful. It is used in social science, psychology and other survey types of research. This functions performs Garrett Ranking of up to 15 ranks.

garrett_ranking(data, num_rank, ranking = NULL, m_rank = c(2:15))

Arguments

data

The data for the Garrett Ranking, must be a data.frame.

num_rank

A vector representing the number of ranks applied to the data. If the data is a five-point Likert-type data, then number of ranks is 5.

ranking

A vector of list representing the ranks applied to the data. If not available, positional ranks are applied.

m_rank

The scope of the ranking methods which is between 2 and 15.

Value

A list with the following components:

RII

Relative importance index.

Garrett ranked data

Table of data ranked using Garrett mean score.

Garrett value

Table of ranking Garrett values

Examples

library(readr)
garrett_data <- data.frame(garrett_data)
ranking <- c("Serious constraint", "Constraint",
"Not certain it is a constraint", "Not a constraint",
"Not a serious constraint")

## ranking is supplied
garrett_ranking(garrett_data, 5, ranking)
#> New names:
#> • `` -> `...1`
#> • `V1` -> `V1...2`
#> • `V1` -> `V1...8`
#> • `` -> `...9`
#> $`Garrett value`
#> # A tibble: 5 × 4
#>   Number `Garrett point` `Garrett index` `Garrett value`
#>    <dbl>           <dbl>           <dbl>           <dbl>
#> 1      1            3.33              15              85
#> 2      2           10                 25              75
#> 3      3           16.7               31              69
#> 4      4           23.3               36              64
#> 5      5           30                 40              60
#> 
#> $`Garrett ranked data`
#>    S/No Description Serious constraint Constraint
#> 1     2          S2                  5          3
#> 2     9          S9                  7          6
#> 3    15         S15                  7          6
#> 4     5          S5                 10          2
#> 5    11         S11                 10          2
#> 6     4          S4                  4          4
#> 7    10         S10                  4          4
#> 8     3          S3                  1          2
#> 9     1          S1                  0          0
#> 10    6          S6                  0          4
#> 11   12         S12                  0          4
#> 12    7          S7                  0          2
#> 13   13         S13                  0          2
#> 14    8          S8                  0          0
#> 15   14         S14                  0          0
#>    Not certain it is a constraint Not a constraint Not a serious constraint
#> 1                               2                2                        1
#> 2                               0                5                        1
#> 3                               0                5                        1
#> 4                               8                5                        0
#> 5                               8                5                        0
#> 6                               6                7                        3
#> 7                               6                7                        3
#> 8                               5                5                        1
#> 9                               2                1                        0
#> 10                              6                5                        6
#> 11                              6                5                        6
#> 12                              0                2                        2
#> 13                              0                2                        2
#> 14                              5                2                       17
#> 15                              5                2                       17
#>    Total Mean Total Garrett Score Mean Garrett score Total Item score
#> 1     13  8.2                 976                 75               48
#> 2     19  4.5                1425                 75               70
#> 3     19  4.5                1425                 75               70
#> 4     25  3.4                1872                 75               92
#> 5     25  3.4                1872                 75               92
#> 6     24  3.3                1682                 70               71
#> 7     24  3.3                1682                 70               71
#> 8     14  6.0                 960                 69               39
#> 9      3 14.8                 202                 67                8
#> 10    21  4.0                1394                 66               50
#> 11    21  4.0                1394                 66               50
#> 12     6  7.0                 398                 66               14
#> 13     6  7.0                 398                 66               14
#> 14    24  1.9                1493                 62               36
#> 15    24  1.9                1493                 62               36
#>    Relative importance index Rank
#> 1                      0.331    1
#> 2                      0.483    2
#> 3                      0.483    3
#> 4                      0.634    4
#> 5                      0.634    5
#> 6                      0.490    6
#> 7                      0.490    7
#> 8                      0.269    8
#> 9                      0.055    9
#> 10                     0.345   10
#> 11                     0.345   11
#> 12                     0.097   12
#> 13                     0.097   13
#> 14                     0.248   14
#> 15                     0.248   15
#> 
#> $RII
#>    V1 V2 V3 V4 V5
#> 1   0  0  6  2  0
#> 2  25 12  6  4  1
#> 3   5  8 15 10  1
#> 4  20 16 18 14  3
#> 5  50  8 24 10  0
#> 6   0 16 18 10  6
#> 7   0  8  0  4  2
#> 8   0  0 15  4 17
#> 9  35 24  0 10  1
#> 10 20 16 18 14  3
#> 11 50  8 24 10  0
#> 12  0 16 18 10  6
#> 13  0  8  0  4  2
#> 14  0  0 15  4 17
#> 15 35 24  0 10  1
#> 

# ranking not supplied
garrett_ranking(garrett_data, 5)
#> New names:
#> • `` -> `...1`
#> • `V1` -> `V1...2`
#> • `V1` -> `V1...8`
#> • `` -> `...9`
#> $`Garrett value`
#> # A tibble: 5 × 4
#>   Number `Garrett point` `Garrett index` `Garrett value`
#>    <dbl>           <dbl>           <dbl>           <dbl>
#> 1      1            3.33              15              85
#> 2      2           10                 25              75
#> 3      3           16.7               31              69
#> 4      4           23.3               36              64
#> 5      5           30                 40              60
#> 
#> $`Garrett ranked data`
#>    S/No Description 1st Rank 2nd Rank 3rd Rank 4th Rank 5th Rank Total Mean
#> 1     2          S2        5        3        2        2        1    13  8.2
#> 2     9          S9        7        6        0        5        1    19  4.5
#> 3    15         S15        7        6        0        5        1    19  4.5
#> 4     5          S5       10        2        8        5        0    25  3.4
#> 5    11         S11       10        2        8        5        0    25  3.4
#> 6     4          S4        4        4        6        7        3    24  3.3
#> 7    10         S10        4        4        6        7        3    24  3.3
#> 8     3          S3        1        2        5        5        1    14  6.0
#> 9     1          S1        0        0        2        1        0     3 14.8
#> 10    6          S6        0        4        6        5        6    21  4.0
#> 11   12         S12        0        4        6        5        6    21  4.0
#> 12    7          S7        0        2        0        2        2     6  7.0
#> 13   13         S13        0        2        0        2        2     6  7.0
#> 14    8          S8        0        0        5        2       17    24  1.9
#> 15   14         S14        0        0        5        2       17    24  1.9
#>    Total Garrett Score Mean Garrett score Total Item score
#> 1                  976                 75               48
#> 2                 1425                 75               70
#> 3                 1425                 75               70
#> 4                 1872                 75               92
#> 5                 1872                 75               92
#> 6                 1682                 70               71
#> 7                 1682                 70               71
#> 8                  960                 69               39
#> 9                  202                 67                8
#> 10                1394                 66               50
#> 11                1394                 66               50
#> 12                 398                 66               14
#> 13                 398                 66               14
#> 14                1493                 62               36
#> 15                1493                 62               36
#>    Relative importance index Rank
#> 1                      0.331    1
#> 2                      0.483    2
#> 3                      0.483    3
#> 4                      0.634    4
#> 5                      0.634    5
#> 6                      0.490    6
#> 7                      0.490    7
#> 8                      0.269    8
#> 9                      0.055    9
#> 10                     0.345   10
#> 11                     0.345   11
#> 12                     0.097   12
#> 13                     0.097   13
#> 14                     0.248   14
#> 15                     0.248   15
#> 
#> $RII
#>    V1 V2 V3 V4 V5
#> 1   0  0  6  2  0
#> 2  25 12  6  4  1
#> 3   5  8 15 10  1
#> 4  20 16 18 14  3
#> 5  50  8 24 10  0
#> 6   0 16 18 10  6
#> 7   0  8  0  4  2
#> 8   0  0 15  4 17
#> 9  35 24  0 10  1
#> 10 20 16 18 14  3
#> 11 50  8 24 10  0
#> 12  0 16 18 10  6
#> 13  0  8  0  4  2
#> 14  0  0 15  4 17
#> 15 35 24  0 10  1
#> 

# you can rank subset of the data
garrett_ranking(garrett_data, 8)
#> New names:
#> • `` -> `...1`
#> • `V1` -> `V1...2`
#> • `V1` -> `V1...11`
#> • `` -> `...12`
#> $`Garrett value`
#> # A tibble: 8 × 4
#>   Number `Garrett point` `Garrett index` `Garrett value`
#>    <dbl>           <dbl>           <dbl>           <dbl>
#> 1      1            3.33              15              85
#> 2      2           10                 25              75
#> 3      3           16.7               31              69
#> 4      4           23.3               36              64
#> 5      5           30                 40              60
#> 6      6           36.7               43              57
#> 7      7           43.3               47              53
#> 8      8           50                 50              50
#> 
#> $`Garrett ranked data`
#>    S/No Description 1st Rank 2nd Rank 3rd Rank 4th Rank 5th Rank 6th Rank
#> 1     7          S7        4        2        2        0        2        0
#> 2    13         S13        4        2        2        0        2        0
#> 3     2          S2        2        0        2        5        3        2
#> 4     9          S9        0        4        4        7        6        0
#> 5    15         S15        0        4        4        7        6        0
#> 6     3          S3        1        3        4        1        2        5
#> 7     5          S5        0        1        0       10        2        8
#> 8    11         S11        0        1        0       10        2        8
#> 9     4          S4        0        1        3        4        4        6
#> 10   10         S10        0        1        3        4        4        6
#> 11    6          S6        0        1        1        0        4        6
#> 12   12         S12        0        1        1        0        4        6
#> 13    1          S1        0        0        0        0        0        2
#> 14    8          S8        1        0        0        0        0        5
#> 15   14         S14        1        0        0        0        0        5
#>    7th Rank 8th Rank Total Mean Total Garrett Score Mean Garrett score
#> 1         2        2    14  7.0                 954                 68
#> 2         2        2    14  7.0                 954                 68
#> 3         2        1    17  8.2                1078                 63
#> 4         5        1    27  4.5                1699                 63
#> 5         5        1    27  4.5                1699                 63
#> 6         5        1    22  6.0                1370                 62
#> 7         5        0    26  3.4                1556                 60
#> 8         5        0    26  3.4                1556                 60
#> 9         7        3    28  3.3                1641                 59
#> 10        7        3    28  3.3                1641                 59
#> 11        5        6    23  4.0                1291                 56
#> 12        5        6    23  4.0                1291                 56
#> 13        1        0     3 14.8                 167                 56
#> 14        2       17    25  1.9                1326                 53
#> 15        2       17    25  1.9                1326                 53
#>    Total Item score Relative importance index Rank
#> 1                72                     0.310    1
#> 2                72                     0.310    2
#> 3                76                     0.328    3
#> 4               122                     0.526    4
#> 5               122                     0.526    5
#> 6                92                     0.397    6
#> 7                99                     0.427    7
#> 8                99                     0.427    8
#> 9                96                     0.414    9
#> 10               96                     0.414   10
#> 11               63                     0.272   11
#> 12               63                     0.272   12
#> 13                8                     0.034   13
#> 14               44                     0.190   14
#> 15               44                     0.190   15
#> 
#> $RII
#>    V1 V2 V3 V4 V5 V6 V7 V8
#> 1   0  0  0  0  0  6  2  0
#> 2  16  0 12 25 12  6  4  1
#> 3   8 21 24  5  8 15 10  1
#> 4   0  7 18 20 16 18 14  3
#> 5   0  7  0 50  8 24 10  0
#> 6   0  7  6  0 16 18 10  6
#> 7  32 14 12  0  8  0  4  2
#> 8   8  0  0  0  0 15  4 17
#> 9   0 28 24 35 24  0 10  1
#> 10  0  7 18 20 16 18 14  3
#> 11  0  7  0 50  8 24 10  0
#> 12  0  7  6  0 16 18 10  6
#> 13 32 14 12  0  8  0  4  2
#> 14  8  0  0  0  0 15  4 17
#> 15  0 28 24 35 24  0 10  1
#> 

garrett_ranking(garrett_data, 4)
#> New names:
#> • `` -> `...1`
#> • `V1` -> `V1...2`
#> • `V1` -> `V1...7`
#> • `` -> `...8`
#> $`Garrett value`
#> # A tibble: 4 × 4
#>   Number `Garrett point` `Garrett index` `Garrett value`
#>    <dbl>           <dbl>           <dbl>           <dbl>
#> 1      1            3.33              15              85
#> 2      2           10                 25              75
#> 3      3           16.7               31              69
#> 4      4           23.3               36              64
#> 
#> $`Garrett ranked data`
#>    S/No Description 1st Rank 2nd Rank 3rd Rank 4th Rank Total Mean
#> 1     9          S9        6        0        5        1    12  4.5
#> 2    15         S15        6        0        5        1    12  4.5
#> 3     2          S2        3        2        2        1     8  8.2
#> 4     5          S5        2        8        5        0    15  3.4
#> 5    11         S11        2        8        5        0    15  3.4
#> 6     3          S3        2        5        5        1    13  6.0
#> 7     4          S4        4        6        7        3    20  3.3
#> 8    10         S10        4        6        7        3    20  3.3
#> 9     1          S1        0        2        1        0     3 14.8
#> 10    7          S7        2        0        2        2     6  7.0
#> 11   13         S13        2        0        2        2     6  7.0
#> 12    6          S6        4        6        5        6    21  4.0
#> 13   12         S12        4        6        5        6    21  4.0
#> 14    8          S8        0        5        2       17    24  1.9
#> 15   14         S14        0        5        2       17    24  1.9
#>    Total Garrett Score Mean Garrett score Total Item score
#> 1                  919                 77               35
#> 2                  919                 77               35
#> 3                  607                 76               23
#> 4                 1115                 74               42
#> 5                 1115                 74               42
#> 6                  954                 73               34
#> 7                 1465                 73               51
#> 8                 1465                 73               51
#> 9                  219                 73                8
#> 10                 436                 73               14
#> 11                 436                 73               14
#> 12                1519                 72               50
#> 13                1519                 72               50
#> 14                1601                 67               36
#> 15                1601                 67               36
#>    Relative importance index Rank
#> 1                      0.302    1
#> 2                      0.302    2
#> 3                      0.198    3
#> 4                      0.362    4
#> 5                      0.362    5
#> 6                      0.293    6
#> 7                      0.440    7
#> 8                      0.440    8
#> 9                      0.069    9
#> 10                     0.121   10
#> 11                     0.121   11
#> 12                     0.431   12
#> 13                     0.431   13
#> 14                     0.310   14
#> 15                     0.310   15
#> 
#> $RII
#>    V1 V2 V3 V4
#> 1   0  6  2  0
#> 2  12  6  4  1
#> 3   8 15 10  1
#> 4  16 18 14  3
#> 5   8 24 10  0
#> 6  16 18 10  6
#> 7   8  0  4  2
#> 8   0 15  4 17
#> 9  24  0 10  1
#> 10 16 18 14  3
#> 11  8 24 10  0
#> 12 16 18 10  6
#> 13  8  0  4  2
#> 14  0 15  4 17
#> 15 24  0 10  1
#>