Index Construction for estimation of Exposure or Sensitivity

Vulnerability or to be vulnerable means the state or quality of being susceptible to physical or emotional harm, damage, or attack, usually indicating a lack of defense or protection, making someone or some systems more likely to be affected by external factors or threats. Therefore, vulnerability index is a quantitative or standardized framework of such state or quality which then makes comparisons between households, communities or systems possible. The index is made up of three main components: exposure, sensitivity and adaptive capacity. Each component has multiple indicators from wide ranges including social, medical, psychological and various extreme events like floods, drought, earthquakes etc. This function is for conversion of indicators exposure and sensitivity into a vector of index through normalization and weighting. The resulting index from each of the component is then combined via an appropriate model into vulnerability index.

Usage

index_construction(data)

Arguments

data

Data frame of indicators of Exposure or Sensitivity. The data frame must be numeric.

Value

A list with the following components:

Indexed data

dataframe of indices corresponding to the supplied data.

Index

A vector of indices representing the variable of interest, either Exposure or Sensitivity.

Examples

library(readr)
garrett_data <- data.frame(garrett_data)
index_construction(garrett_data)
#> $`Indexed data`
#>      S1   S2   S3   S4   S5   S6   S7   S8   S9   S10  S11  S12  S13  S14   S15
#> 1  0.23 0.00 0.00 0.00 0.00 0.00 0.00 0.00 0.00 0.000 0.00 0.00 0.00 0.00 0.000
#> 2  0.00 0.11 0.00 0.00 0.00 0.00 0.00 0.00 0.00 0.000 0.00 0.00 0.00 0.00 0.298
#> 3  0.39 0.00 0.25 0.00 0.00 0.00 0.00 0.00 0.00 0.000 0.00 0.00 0.00 0.00 0.000
#> 4  0.00 0.25 0.00 0.11 0.00 0.00 0.00 0.00 0.00 0.000 0.00 0.00 0.00 0.00 0.000
#> 5  0.00 0.00 0.17 0.00 0.19 0.00 0.00 0.00 0.00 0.000 0.00 0.00 0.00 0.00 0.000
#> 6  0.00 0.00 0.00 0.00 0.00 0.11 0.00 0.00 0.00 0.000 0.00 0.00 0.00 0.00 0.000
#> 7  0.00 0.00 0.00 0.00 0.11 0.00 0.00 0.00 0.00 0.000 0.00 0.00 0.00 0.00 0.000
#> 8  0.00 0.00 0.00 0.00 0.00 0.17 0.00 0.00 0.00 0.000 0.00 0.00 0.00 0.00 0.000
#> 9  0.00 0.00 0.00 0.00 0.00 0.00 0.21 0.00 0.16 0.000 0.00 0.00 0.00 0.00 0.000
#> 10 0.00 0.00 0.00 0.00 0.00 0.00 0.00 0.17 0.00 0.069 0.00 0.00 0.00 0.00 0.000
#> 11 0.00 0.00 0.00 0.00 0.00 0.00 0.00 0.00 0.25 0.000 0.11 0.00 0.00 0.00 0.000
#> 12 0.00 0.00 0.00 0.00 0.00 0.00 0.00 0.00 0.00 0.275 0.00 0.00 0.00 0.00 0.000
#> 13 0.00 0.00 0.00 0.00 0.00 0.00 0.00 0.00 0.00 0.000 0.17 0.00 0.24 0.00 0.000
#> 14 0.00 0.00 0.00 0.00 0.00 0.00 0.00 0.00 0.00 0.000 0.00 0.11 0.00 0.11 0.000
#> 15 0.00 0.00 0.00 0.00 0.00 0.00 0.00 0.00 0.00 0.000 0.00 0.00 0.11 0.00 0.069
#> 16 0.28 0.00 0.00 0.00 0.00 0.00 0.00 0.00 0.00 0.000 0.00 0.00 0.00 0.00 0.000
#> 17 0.00 0.23 0.00 0.00 0.00 0.00 0.00 0.00 0.00 0.000 0.00 0.00 0.00 0.00 0.275
#> 18 0.44 0.00 0.11 0.00 0.00 0.00 0.00 0.00 0.00 0.000 0.00 0.00 0.00 0.00 0.000
#> 19 0.00 0.38 0.00 0.18 0.00 0.00 0.00 0.00 0.00 0.000 0.00 0.00 0.00 0.00 0.000
#> 20 0.00 0.00 0.21 0.00 0.16 0.00 0.00 0.00 0.00 0.000 0.00 0.00 0.00 0.00 0.000
#> 21 0.00 0.00 0.00 0.11 0.00 0.11 0.00 0.00 0.00 0.000 0.00 0.00 0.00 0.00 0.000
#> 22 0.00 0.00 0.00 0.00 0.21 0.00 0.14 0.00 0.00 0.000 0.00 0.00 0.00 0.00 0.000
#> 23 0.00 0.00 0.00 0.00 0.00 0.21 0.00 0.00 0.00 0.000 0.00 0.00 0.00 0.00 0.000
#> 24 0.00 0.00 0.00 0.00 0.00 0.00 0.35 0.00 0.16 0.000 0.00 0.00 0.00 0.00 0.000
#> 25 0.00 0.00 0.00 0.00 0.00 0.00 0.00 0.00 0.00 0.160 0.00 0.00 0.00 0.00 0.000
#> 26 0.00 0.00 0.00 0.00 0.00 0.00 0.00 0.00 0.27 0.000 0.11 0.00 0.00 0.00 0.000
#> 27 0.00 0.00 0.00 0.00 0.00 0.00 0.00 0.00 0.00 0.212 0.00 0.00 0.00 0.00 0.000
#> 28 0.00 0.00 0.00 0.00 0.00 0.00 0.00 0.00 0.00 0.000 0.25 0.00 0.22 0.00 0.000
#> 29 0.00 0.00 0.00 0.00 0.00 0.00 0.00 0.00 0.00 0.000 0.00 0.44 0.00 0.23 0.000
#> 
#> $Index
#>  [1] 0.0152 0.0271 0.0432 0.0237 0.0241 0.0073 0.0071 0.0112 0.0248 0.0158
#> [11] 0.0237 0.0183 0.0271 0.0143 0.0117 0.0187 0.0335 0.0368 0.0372 0.0248
#> [21] 0.0143 0.0233 0.0142 0.0342 0.0106 0.0256 0.0142 0.0310 0.0448
#>